Fix bug: Error in pulling stock prices
As it turns out Beautiful Soup wasn't pulling the correct data as I specified finding an element with the Trsdu(0.3) class. More than one element has this class, so it identified the wrong element as the price. 

This commit fixes this by adding all the classes applied to the price element, so the correct HTML element is identified.